marc utility with split marc records
