+++ src/main/java/org/folio/rest/utility/MarcUtility.java
@@ -1,27 +1,15 @@
 package org.folio.rest.utility;
 
-import static java.util.regex.Pattern.CASE_INSENSITIVE;
-import static java.util.regex.Pattern.MULTILINE;
-import static java.util.regex.Pattern.compile;
-
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.regex.Pattern;
-
-import org.marc4j.MarcException;
-import org.marc4j.MarcJsonReader;
-import org.marc4j.MarcJsonWriter;
-import org.marc4j.MarcStreamReader;
-import org.marc4j.marc.Record;
-import org.marc4j.marc.Subfield;
-import org.marc4j.marc.VariableField;
-import org.marc4j.marc.impl.SubfieldImpl;
+import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import com.fasterxml.jackson.core.JsonParser;
@@ -33,25 +21,22 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 
-/**
- * A collection of utilities methods intended to be directly used by scripts
- * called by the scripting engine.
- */
-public class ScriptEngineUtility {
+import org.marc4j.MarcException;
+import org.marc4j.MarcJsonReader;
+import org.marc4j.MarcJsonWriter;
+import org.marc4j.MarcStreamReader;
+import org.marc4j.marc.Record;
+import org.marc4j.marc.Subfield;
+import org.marc4j.marc.VariableField;
+import org.marc4j.marc.impl.SubfieldImpl;
 
-  private static final String EMAIL_REGEX = "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$";
-  private static final String PHONE_REGEX = "^[\\+]?[(]?[0-9]{3}[)]?[-\\s\\.]?[0-9]{3}[-\\s\\.]?[0-9]{4,6}$";
-  private static final String URL_REGEX = "(http(s)?:\\\\\\/\\\\\\/.)?(www\\.)?[-a-zA-Z0-9@:%._\\\\\\+~#=]{2,256}\\\\\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_\\\\\\+.~#?&//=]*)";
-
-  private static final Pattern EMAIL_PATTERN = compile(EMAIL_REGEX, CASE_INSENSITIVE | MULTILINE);
-  private static final Pattern PHONE_PATTERN = compile(PHONE_REGEX, CASE_INSENSITIVE | MULTILINE);
-  private static final Pattern URL_PATTERN = compile(URL_REGEX, CASE_INSENSITIVE | MULTILINE);
-
-  private static final ObjectMapper mapper = new ObjectMapper();
+public class MarcUtility {
 
   private static final Charset DEFAULT_CHARSET = StandardCharsets.UTF_8;
 
-  public ScriptEngineUtility() {
+  private static final ObjectMapper mapper = new ObjectMapper();
+
+  public MarcUtility() {
     mapper.setSerializationInclusion(Include.NON_NULL);
     SimpleModule module = new SimpleModule();
     module.addDeserializer(Subfield.class, new JsonDeserializer<Subfield>() {
@@ -66,76 +51,24 @@ public class ScriptEngineUtility {
     mapper.registerModule(module);
   }
 
-  /**
-   * Check if a given string is a proper e-mail address.
-   *
-   * @param string The string to check.
-   *
-   * @return TRUE if matched, FALSE otherwise.
-   */
-  public boolean isEmail(String string) {
-    return EMAIL_PATTERN.matcher(string).find();
-  }
-
-  /**
-   * Check if a given string might be an e-mail.
-   *
-   * @param string The string to check.
-   *
-   * @return TRUE if matched, FALSE otherwise.
-   */
-  public boolean isEmailLike(String string) {
-    return isEmail(string) || (string == null ? false : string.toLowerCase().indexOf("@") != -1);
-  }
-
-  /**
-   * Check if a given string is a phone number.
-   *
-   * @param string The string to check.
-   *
-   * @return TRUE if matched, FALSE otherwise.
-   */
-  public boolean isPhone(String string) {
-    return PHONE_PATTERN.matcher(string).find();
-  }
-
-  /**
-   * Check if a given string might be a URL.
-   *
-   * @param string The string to check.
-   *
-   * @return TRUE if matched, FALSE otherwise.
-   */
-  public boolean isURLLike(String string) {
-    if (string == null) {
-      return false;
+  public List<String> splitRawMarcToMarcJsonRecords(String rawMarc) {
+    List<Record> records = new ArrayList<>();
+    try (InputStream in = new ByteArrayInputStream(rawMarc.getBytes(DEFAULT_CHARSET))) {
+      final MarcStreamReader reader = new MarcStreamReader(in, DEFAULT_CHARSET.name());
+      if (reader.hasNext()) {
+        records.add(reader.next());
+      }
+    } catch (IOException e) {
+      // TODO: do something in case of exception
+      System.out.println(e.getMessage());
+    } catch (final MarcException e) {
+      // TODO: do something in case of exception
+      System.out.println(e.getMessage());
     }
 
-    boolean isLikeUrl = string.toLowerCase().indexOf("www.") != -1;
-
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".org") != -1;
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".edu") != -1;
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".net") != -1;
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".us") != -1;
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".io") != -1;
-    isLikeUrl = isLikeUrl || string.toLowerCase().indexOf(".co") != -1;
-
-    return (isValidUrl(string) || isLikeUrl) && !isEmailLike(string);
-  }
-
-  /**
-   * Check if a given string is a valid URL.
-   *
-   * @param string The string to check.
-   *
-   * @return TRUE if matched, FALSE otherwise.
-   */
-  public boolean isValidUrl(String string) {
-    if (string == null) {
-      return false;
-    }
-
-    return URL_PATTERN.matcher(string).find();
+    return records.stream()
+      .map(this::recordToJson)
+      .collect(Collectors.toList());
   }
 
   /**
